requirements.txt
aenum==2.0.8
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
enum34==1.1.6
fake-useragent==0.1.10
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipaddress==1.0.22
lxml==4.2.1
parsel==1.4.0
peewee==3.5.2
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pymitter==0.2.3
PyMySQL==0.9.2
pyOpenSSL==17.5.0
PyPika==0.14.8
python-dotenv==0.8.2
pyzk==0.6
queuelib==1.5.0
requests==2.18.4
service-identity==17.0.0
six==1.11.0
tqdm==4.23.0
Twisted==18.4.0
urllib3==1.22
virtualenv==16.0.0
w3lib==1.19.0
zope.interface==4.5.0
